Add note to build docs on limiting cloned size via --depth
When cloning, I had assumed the repo's files were just huge for some reason until I saw in https://github.com/be5invis/Iosevka/issues/1122 that it was really just the commit history

This publicizes that more and may save others some bandwidth.

-----

Repeat of incorrectly submitted #1395 

Thanks for the beautiful font! I switched over to it a few days ago and it's been fantastic. Simply the most I've ever liked a font, and the customization options are superb.