Adjust variant assignments based on original fonts
I was able to find a copy of all fonts emulated under Iosevka, including "Fixed" which I presume is what SS11 is based on, except for Pragmata Pro which I referenced their specimen for, and found that this is how each of these fonts handle the characters in question.

Additionally, PARTIAL DIFFERENTIAL (`U+2202`) is made curly by default, with only `ss02` using the straight style.