Improve legibility of quadruple arrows
Double arrows have a minimum gap size equal to the bar size;
Triple arrows have a minimum gap size of 75% of the bar size;
Quadruple arrows will now have a minimum gap size of 62.5% of the bar size.

This follows a trend of halving the distance between 100% and 50% with every additional bar. If there were a quintuple arrow in Unicode, then the gap should be 56.25% of the bar size, and so on, but never reaching 50%.